Aded many things in order for the jwt to work
