+++ belatro-backend/src/main/resources/.env
@@ -1,0 +1,6 @@
+MONGO_USERNAME=admin
+MONGO_PASSWORD=QpE4FMId6oVP8GwJ
+REDIS_LINK=redis-17569.c293.eu-central-1-1.ec2.redns.redis-cloud.com
+REDIS_PORT=17569
+REDIS_USER=default
+REDIS_PASS=VX3Ywm5KCs9zxswTjCRjxyMvsKtrzfBr
